SIM-863 UPDATE: Add counting for easy rules, refine unit tests

+++ ui/src/main/java/eu/jsparrow/ui/preview/RefactoringPreviewWizardPage.java
@@ -78,7 +78,7 @@ public class RefactoringPreviewWizardPage extends WizardPage {
 			AbstractRefactoringRule<? extends AbstractASTRewriteASTVisitor> rule) {
 		super(rule.getName());
 		setTitle(rule.getName());
-		setDescription(rule.getDescription() +" Times rule was applied: " + RuleApplicationCounter.get(rule)); //$NON-NLS-1$
+		setDescription(rule.getDescription() +" Times rule was applied: " + RuleApplicationCounter.get(rule).toInt()); //$NON-NLS-1$
 
 		this.changesForRule = changesForRule;
 		this.rule = rule;
